src/client/app/shared/directives/colorcode-clia.directive.spec.ts
import { TestBed, ComponentFixture, async } from '@angular/core/testing';
import { Component, DebugElement } from '@angular/core';
import { By } from '@angular/platform-browser';

import { ColorCodeCliaDirective } from './colorcode-clia.directive';

export function main() {
    describe('Pipe: Default', () => {
        let testHost: TestHostComponent;
        let fixture: ComponentFixture<TestHostComponent>;
        let testEl: DebugElement;

        beforeEach(async(() => {
            TestBed.configureTestingModule({
                declarations: [TestHostComponent, ColorCodeCliaDirective]
            }).compileComponents();
        }));

        beforeEach(() => {
            fixture = TestBed.createComponent(TestHostComponent);
            testHost = fixture.componentInstance;
            testEl = fixture.debugElement.query(By.css('span'));
            fixture.detectChanges();
        });

        it('empty value adds no color class', () => {
            expect(testEl.nativeElement.classList.contains('text-success-light')).toBe(false);
            expect(testEl.nativeElement.classList.contains('text-danger-light')).toBe(false);
            expect(testEl.nativeElement.classList.contains('text-purple-light')).toBe(false);
        });

        it('"PASSED" adds "text-success-light" color class', () => {
            testHost.testStatus = 'PASSED';
            fixture.detectChanges();
            expect(testEl.nativeElement.classList.contains('text-success-light')).toBe(true);
            expect(testEl.nativeElement.classList.contains('text-danger-light')).toBe(false);
            expect(testEl.nativeElement.classList.contains('text-purple-light')).toBe(false);
        });

        it('"FAILED" adds "text-danger-light" color class', () => {
            testHost.testStatus = 'FAILED';
            fixture.detectChanges();
            expect(testEl.nativeElement.classList.contains('text-success-light')).toBe(false);
            expect(testEl.nativeElement.classList.contains('text-danger-light')).toBe(true);
            expect(testEl.nativeElement.classList.contains('text-purple-light')).toBe(false);
        });

        it('"PENDING" adds "text-purple-light" color class', () => {
            testHost.testStatus = 'PENDING';
            fixture.detectChanges();
            expect(testEl.nativeElement.classList.contains('text-success-light')).toBe(false);
            expect(testEl.nativeElement.classList.contains('text-danger-light')).toBe(false);
            expect(testEl.nativeElement.classList.contains('text-purple-light')).toBe(true);
        });
    });
}

@Component({
    template: `<span [colorCodeClia]="testStatus">Some Text</span>`
})
class TestHostComponent {
    testStatus: string;
}
